pyproject.toml
[project]
name = "todoproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.4",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
]
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "todoproject.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
